Added search & distance calculation & fragments & conflicts resolved
