+++ app/src/main/java/com/froyo/ridekaro/views/MainActivity.kt
@@ -11,7 +11,6 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-
     }
 }
