--- app/src/main/res/layout/activity_user_home_screen.xml
@@ -1,123 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/parent"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".views.UserHomeScreen">
-
-    <com.google.android.gms.maps.MapView
-        android:id="@+id/googleMap"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toTopOf="@id/currentLocationEdit"
-        app:layout_constraintTop_toTopOf="@id/parent" />
-
-    <ImageView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/green_pin"
-        app:layout_constraintBottom_toBottomOf="@id/googleMap"
-        app:layout_constraintEnd_toEndOf="@id/googleMap"
-        app:layout_constraintStart_toStartOf="@id/googleMap"
-        app:layout_constraintTop_toTopOf="@id/googleMap" />
-
-    <LinearLayout
-        android:id="@+id/currentLocationEdit"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="@dimen/dp_10"
-        android:background="@drawable/edittext_bg"
-        app:layout_constraintBottom_toTopOf="@id/searchLocationEdit">
-
-        <androidx.cardview.widget.CardView
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/dp_40"
-            android:layout_margin="@dimen/dp_15">
-
-            <ImageView
-                android:layout_width="@dimen/dp_15"
-                android:layout_height="@dimen/dp_15"
-                android:layout_margin="@dimen/dp_12"
-                android:src="@drawable/green_circle" />
-
-            <TextView
-                android:id="@+id/tvCurrentAddress"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginStart="@dimen/dp_40"
-                android:layout_marginTop="@dimen/dp_5"
-                android:layout_marginEnd="@dimen/dp_10"
-                android:fontFamily="@font/opensansregular"
-                android:text="@string/current_location"
-                android:textColor="@color/black"
-                android:textSize="@dimen/sp_17" />
-        </androidx.cardview.widget.CardView>
-
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/searchLocationEdit"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="@dimen/dp_10"
-        android:background="@drawable/edittext_bg"
-        app:layout_constraintBottom_toTopOf="@id/bottomLinearLayout">
-
-        <androidx.cardview.widget.CardView
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/dp_40"
-            android:layout_margin="@dimen/dp_15">
-
-            <ImageView
-                android:layout_width="@dimen/dp_15"
-                android:layout_height="@dimen/dp_15"
-                android:layout_margin="@dimen/dp_12"
-                android:src="@drawable/red_circle" />
-
-            <TextView
-                android:id="@+id/tvEnterDestination"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginStart="@dimen/dp_40"
-                android:layout_marginTop="@dimen/dp_5"
-                android:layout_marginEnd="@dimen/dp_10"
-                android:fontFamily="@font/opensansregular"
-                android:text="@string/enter_destination"
-                android:textColor="@color/black"
-                android:textColorHint="@color/black"
-                android:textSize="@dimen/sp_17" />
-        </androidx.cardview.widget.CardView>
-
-
-    </LinearLayout>
-
-
-    <LinearLayout
-        android:id="@+id/bottomLinearLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@color/BgDarkGray"
-        android:gravity="center"
-        android:orientation="vertical"
-        app:layout_constraintBottom_toBottomOf="@id/parent">
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/bike" />
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="@dimen/dp_5"
-            android:fontFamily="@font/opensansregular"
-            android:text="@string/ride"
-            android:textColor="@color/yellow"
-            android:textSize="@dimen/sp_13" />
-
-    </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
